+++ src/gameserver/core/framework/BundleLoader/IBundle.ts
@@ -1,0 +1,6 @@
+import { IModule } from './IModule';
+
+export type IBundle = {
+    name: String,
+    modules: IModule<any>[],
+}