+++ src/gameserver/core/framework/BundleLoader/IBundleLoader.ts
@@ -1,0 +1,5 @@
+import { IBundle } from './IBundle';
+
+export interface IBundleLoader {
+    loadBundle(bundle: IBundle): void;
+}