+++ src/gameserver/core/framework/BundleLoader/IModule.ts
@@ -1,0 +1,4 @@
+export interface IModule<T extends String> {
+    type: T,
+    name: String,
+}