+++ src/gameserver/core/framework/BundleLoader/IModuleLoaderStrategy.ts
@@ -1,0 +1,6 @@
+import { IModule } from './IModule';
+
+export interface IModuleLoaderStrategy<T extends String, M extends IModule<T>> {
+    readonly type: T;
+    load(module: M): void;
+}