+++ src/gameserver/core/framework/BundleLoader/index.ts
@@ -1,0 +1,5 @@
+export { BundleLoader } from './impl/BundleLoader';
+export { IBundle } from './IBundle';
+export { IBundleLoader } from './IBundleLoader';
+export { IModule } from './IModule';
+export { IModuleLoaderStrategy } from './IModuleLoaderStrategy';