+++ src/gameserver/core/framework/ModuleLoaderStrategies/ServiceModuleLoaderStrategy/index.ts
@@ -1,0 +1,2 @@
+export { IServiceModule } from './IServiceModule';
+export { ServiceModuleLoaderStrategy } from './ServiceModuleLoaderStrategy';