+++ src/gameserver/core/framework/ModuleLoaderStrategies/SystemModuleLoaderStrategy/index.ts
@@ -1,0 +1,2 @@
+export { ISystemModule } from './ISystemModule';
+export { SystemModuleLoaderStrategy } from './SystemModuleLoaderStrategy';