+++ src/gameserver/core/framework/ModuleLoaderStrategies/index.ts
@@ -1,0 +1,4 @@
+export * as moduleTypes from './moduleTypes';
+
+export { ServiceModuleLoaderStrategy } from './ServiceModuleLoaderStrategy';
+export { SystemModuleLoaderStrategy } from './SystemModuleLoaderStrategy';