+++ src/gameserver/core/framework/ModuleLoaderStrategies/moduleTypes.ts
@@ -1,0 +1,7 @@
+import { IServiceModule } from './ServiceModuleLoaderStrategy';
+import { ISystemModule } from './SystemModuleLoaderStrategy';
+
+export {
+    IServiceModule,
+    ISystemModule,
+};