+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased] - <date>
 ### Added
 - Game Server `DependencyLocator`
+- Game Server `BundleLoader`
+- Game Server `SystemModuleLoaderStrategy`
+- Game Server `ServiceModuleLoaderStrategy`
 
 ### Changed
 ### Deprecated